[DEV] bev rasterization algorithm
introduces data types for a new observation format, a rasterized BEV
grid. The grid is formed via a simple raster scan algorithm.
Objects are converted to coordinates on a grid relative to the ego
agent, then have a maximum bounding box computed, which is used to fill
in what is essentially a segmentation map.