[FEATURE] Add Birds-Eye-View observation tensors (#390)
* [DEV] bev rasterization algorithm

introduces data types for a new observation format, a rasterized BEV
grid. The grid is formed via a simple raster scan algorithm.
Objects are converted to coordinates on a grid relative to the ego
agent, then have a maximum bounding box computed, which is used to fill
in what is essentially a segmentation map.

* [DEV] connect bev observations in sim

exposes the bev observations system in the simulator. We iterate over
first over road objects, then over other agents, filling in the
segmenation grid using the grid rasterizer.

* [DEV] expose bev observation tensor in python

* [DEV] expose bev observations in python envs

* [BUILD] add rasterizer.hpp to CMakeLists.txt

---------

Co-authored-by: Ellington Kirby <ellingtonkirby@gmail.com>